backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
alembic==1.13.0
asyncpg==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
websockets==13.0
httpx==0.27.0
pytest==8.3.0
pytest-asyncio==0.24.0
openai>=2.8.0
opencv-python-headless==4.10.0.84
